Added the possibility to add numbers with 3 decimals in tasks/tickets durations (#3)

+++ src/database/migrations/2017_03_09_101753_update_projects_table_add_status_delete_times.php
@@ -29,8 +29,8 @@ class UpdateProjectsTableAddStatusDeleteTimes extends Migration
     {
         Schema::table('projects', function (Blueprint $table) {
             $table->dropColumn('status_id');
-            $table->float('tasks_scheduled_time')->after('color')->nullable();
-            $table->float('tickets_scheduled_time')->after('tasks_scheduled_time')->nullable();
+            $table->double('tasks_scheduled_time', 8, 3)->after('color')->nullable();
+            $table->double('tickets_scheduled_time', 8, 3)->after('tasks_scheduled_time')->nullable();
         });
     }
 }
